Updated the application to use plugin based components over hard-coded ones.
Have begun to implement the interfaces that plug-ins should adhere to in
order for it to work and looked at the navigation sequence (see
master.blade.php).

Potentially is a good start and a good way to go as it will allow for
claims-based navigation links, headers etc whilst still making it
dynamic and easy to use.